+++ lib/internal/streams/readable.js
@@ -22,7 +22,6 @@
 'use strict';
 
 const {
-  ArrayIsArray,
   NumberIsInteger,
   NumberIsNaN,
   ObjectDefineProperties,
@@ -38,7 +37,7 @@ module.exports = Readable;
 Readable.ReadableState = ReadableState;
 
 const EE = require('events');
-const Stream = require('internal/streams/legacy');
+const { Stream, prependListener } = require('internal/streams/legacy');
 const { Buffer } = require('buffer');
 
 let debug = require('internal/util/debuglog').debuglog('stream', (fn) => {
@@ -68,24 +67,6 @@ ObjectSetPrototypeOf(Readable, Stream);
 function nop() {}
 
 const { errorOrDestroy } = destroyImpl;
-
-function prependListener(emitter, event, fn) {
-  // Sadly this is not cacheable as some libraries bundle their own
-  // event emitter implementation with them.
-  if (typeof emitter.prependListener === 'function')
-    return emitter.prependListener(event, fn);
-
-  // This is a hack to make sure that our error handler is attached before any
-  // userland ones.  NEVER DO THIS. This is here only because this code needs
-  // to continue to work with older versions of Node.js that do not include
-  // the prependListener() method. The goal is to eventually remove this hack.
-  if (!emitter._events || !emitter._events[event])
-    emitter.on(event, fn);
-  else if (ArrayIsArray(emitter._events[event]))
-    emitter._events[event].unshift(fn);
-  else
-    emitter._events[event] = [fn, emitter._events[event]];
-}
 
 function ReadableState(options, stream, isDuplex) {
   // Duplex streams are both readable and writable, but share
