+++ lib/stream.js
@@ -36,7 +36,7 @@ const internalBuffer = require('internal/buffer');
 // Lazy loaded
 let promises = null;
 
-const Stream = module.exports = require('internal/streams/legacy');
+const Stream = module.exports = require('internal/streams/legacy').Stream;
 Stream.Readable = require('internal/streams/readable');
 Stream.Writable = require('internal/streams/writable');
 Stream.Duplex = require('internal/streams/duplex');
